Extract ProModal to indepent file
